aggregator/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
requests==2.31.0
pytest==7.4.3
